Suse-36 Extras Not Showing Fix
Set the extras list grid to unsuspend layout after initializing the data.

This appeared to be a caching issue: if cache was disabled you could see the extras fine. But with cache, it would not show. The items would be there, but the heights would not be updated to show it.

Best guess is issues with extjs autocaching store data and not triggering a built in event when loading from cache.
The data load likely triggers a "resumeLayout" or "doLayout" event. But that is skipped when loading from cache.

Setting the Grid to not suspend layout after adding to the items list forces this event to occur.