COM-1503 Fix Order Details action not matching route
The Order Details action ('Deets') was not matching the route as the shipmentNumber parameter was not optional. It was optional until recently.
As this action was really 2 distinct actions it was split into two actions: Deets and PackingSlip.

(cherry picked from commit 5a6626af3afd679c4762388c9cbc58043df284f6)