COM-1596 Fix B2b lists when adding configurable product
Introduced when trying to handle configurable products with minimum quantity limits. My fix also changed the flow a bit. Instead of showing an error when the product is added, stopping the add to list, it now auto updates to the min qty automatically, adds to list, and informs the user. This is the same flow we use on the cart and product page.