FUL-1524 Refactor back office templates used by 'pick-wave-print'
The back office template 'pick-wave-print' is including 'pick-slip' and 'packing-slip' which are top-level templates that extend the 'back-office/base' template.
The base template uses the body-content block which is defined in all these top-level templates. This appears to cause a recursion loop which eventually times out.
These templates will now work similar to the 'order-pick-sheet' template which includes a 'order-pick-sheet-content' template.