COM-2191 Fix customer refresh during guest checkout
When checking out as a guest and registering a new customer account the customer model is not refreshed and causes a later call to apiModel.getContacts() to fail.
Why does this happen? The SDK looks for an 'id' property on the customer model and since it is undefined the URL parameters are not populated.
This sends the request to "/api/commerce/customer/accounts//contacts/" instead of "/api/commerce/customer/accounts/<accountId>/contacts/?userId=<userid>" as defined in the SDK.
Making the logic inside the captureCustomer method grab the right data fixes the refresh issue. It's unclear when this bug was introduced but this issue does not exist in classic eComm.

(cherry picked from commit a86f78c7abf56c2df1d376ffdbc0d0248c98278c)