COM-2133 Include user claims when creating a new shopper account (#17)
When creating a new registered shopper account, we need to include the current (anonymous) user claims. The UserId from the anonymous claims are added to the new registered claims as a PreviousAnonymousId entry on the user claims bag. When fetching an order, CommerceRuntime checks the Order's UserId against the user claims UserId and PreviousAnonymousId. If either matches then the shopper is granted access to their order. If the anonymous claims are not provided and thus the PreviousAnonymousId is blank, Commerce will throw a 403 forbidden when accessing your order if it was created prior to registering.