COM-2191 Return auth ticket from LoginAndTrack (#19)
When registering a guest shopper core theme requires the entire auth ticket in the response (instead of just the customer) as it looks at the CustomerAccount property when refreshing the customer model.

This change looks to have been introduced during the .NET Core migration for Storefront. The source commit appears to be a435b876f43c8dc5a12193cf234bac092aa00e54 from May 13. Testing on classic storefront does not have this bug.