COM-2597 B2B P2: Add Quotes panel on My Account and Seller Account (#272)
As part of B2B phase 2, enhance My Account to show a new Quotes panel with the buyer's quotes in a grid as well as a create quote section. Seller Account is also enhanced with a new Quotes panel to show the seller's quotes as well as a create quote section with an account search dropdown. The current implementation shows all quotes on the tenant but will be enhanced later to only show quotes the buyer or seller has permission to view (backend work). The create quote button is not yet implemented.

This story also changes the default sort behavior for all grids/collections to sort by ascending, descending, and then unsorted. Includes a new grid component and sort icons for use by the B2B grids.

- Add Quotes panel on My Account and Seller Account pages with a quotes grid and create quote section.
- Added account search dropdown on seller page.
- Improve sorting functionality and add icons.
- Make some cosmetic changes.
- Performance: On Seller Account only make single call to fetch a B2B Account per unique account ID in the Quotes collection.

Co-authored-by: RajeshDatir <rajesh.datir@kibocommerce.com>
Co-authored-by: BhushanSharma <Bhushan.Sharma@kibocommerce.com>