COM-2597 B2B P2: Add Quotes panel on My Account and Seller Account (#43)
As part of B2B phase 2, enhance My Account and Seller Account to show a Quotes panel. UI details can be found in the Mozu.CoreTheme commit.

An important part of this story is enhancing the SDK to support making requests as a seller (i.e., Sales Rep). The seller does not have a valid shopper (buyer) user claims so we need the request to use the admin user (seller) user claims. This is done by setting the user scope type header "x-vol-scope-type" which instructs Reverse Proxy which user claims to use.

- Add Quote API methods to the SDK.
- Update PageContext to support setting the user scope type value and passing it down to the SDK. Only affects the current request.
- Set the user scope type header when viewing Seller Account.
- Update Mozu.CommerceRuntime contracts to the latest B2B phase 2 branch version.
- Update Mozu.CoreTheme to the latest B2B phase 2 branch version.