COM-2600 B2B P2: Add search by address modal on Seller Account (#280)
Add a new search by address modal for use by sellers on the Seller Account page. This modal includes a contacts grid with filters and searches for B2B accounts by contacts. Once an account is selected it can be used to create a Quote for the account tied to that contact. The create quote button functionality will be added in another commit.

Co-authored-by: Hrishikesh <hrishikesh>