COM-2730 B2B P2: Add child account modal (#285)
As part of B2B phase 2, buyers need to be able to create child B2B accounts in the account hierarchy. To support this add a new child account modal which has a form with the necessary inputs to create a new B2B account in the account hierarchy.
Includes validation, error handling, and API integration.

Co-authored-by: Hrishikesh <hrishikesh>