COM3256: B2B P2 - Email & print Quote template issues
Removed the “Quote Name” text before the actual quote name in the title/header.

Add a Quote Number field after Account Name but before Quote Name. Map it to Number on the Quote.

Rename the label for Submitted Date to Created Date. Map it to AuditInfo.CreateDate on the Quote instead of SubmittedDate (which is getting removed).

Rename the label for Submitted By to Created By. This field should do a lookup of the UserId field on the Quote. This should always match the buyer's user ID which should be one of the user's on the account.
Add a colon to Shipping Method to match the other fields. Should say Shipping Method:.