COM-2980 B2B P2: Edit Quote page API integration (part 2) (#292)
The Edit Quote page will be used by buyers and sellers when creating or editing Quotes during the negotiation process. This commit covers API integration (part 2) for the following areas and fields. Minus the shipping section this page is functionally complete.

1) Quote comments - User is able to add and view comments.
2) Quote audit history - User can see Audit History.
3) Shipping - Only shipping address is integrated right now. Only a buyer with an admin role can create or edit the destination address.
4) Quote actions - Enabling and disabling quote buttons based on the quote state.

This commit also includes various bug fixes and tweaks.

Co-authored-by: Sagar Raut <sagar.raut@kibocommerce.com>