COM-3161, COM-3392 B2B P2: Add View Quote page (#301)
As part of B2B phase 2, a buyer needs to be able to view a quote in read-only mode. This is especially true if the user does not have edit permissions but still wants to view the quote. Certain quote statuses like Completed and Expired are always viewed read-only and do not allow edits.

COM-3161:
- Add View Quote page and templates. Move Edit Quote templates into a subfolder. Share common templates between pages.

COM-3392:
- Allow a buyer to checkout on the View Quote page if they are a purchaser.

Co-authored-by: Onkar Patil <onkar.patil@kibocommerce.com>