COM-3161 B2B P2: Add View Quote page (#61)
As part of B2B phase 2, a buyer needs to be able to view a quote in read-only mode. This is especially true if the user does not have edit permissions but still wants to view the quote. Certain quote statuses like Completed and Expired are always viewed read-only and do not allow edits. See the related commit in Mozu.CoreTheme.

- Add View Quote action to My Account and Seller Account controllers.
- Update Mozu.CoreTheme to the latest B2B phase 2 branch version.