Fix quote page size to something tolerable
One of these should fix the quote page size to be 20 instead of 5. I think the change in b2b-account/quotes.js for QuotesGridCollectionModel is the main one that works, but I haven't extensively tested or see where all these collections are used. Also not sure why you have to use defaults instead of just setting pageSize directly on the grid.