COM-3259 B2B P2: Add Quote emails (#307)
Add new templates for Quote emails. In Review emails are sent to sellers and link to the b2B account page on Admin. Ready for Checkout emails are sent to buyers and link to the View Quote page. Expired emails are sent to both buyers and sellers can can link to either Admin or Storefront based on the recipient.