COM-2593 B2B P2: Enhance Quick Order to display the cart (#58)
As part of B2B phase 2, the Quick Order page is being enhanced to support displaying the contents of the buyer's cart. This includes the ability to create a Quote from the current cart by using the Initiate Quote button. See the related commit in Mozu.CoreTheme.

The cart view was refactored to be shared between the Cart page and Quick Order page.

- Add SDK methods for Cart Details and Create Quote from Cart.
- Refactor the Cart and Quick Order APIs to both support the cart.
- Add a GetCart API to support getting items in the cart.
- Add quick order support for xpress-paypal.
- Update Mozu.CoreTheme to the latest B2B phase 2 branch version.