COM-3580 B2B P2: Add View Quote action to the View B2B Account page (#316)
Add support for the View Quote action on the View B2B Account page quotes grid. This is necessary to allow buyers to view quotes on a child/descendant account as they can not be viewed from the My Account quotes grid.
Also remove View Account option from the account hierarchy UI if the account is the user's own account.