COM-3603 B2B P2: Fix My Account to not Get Quotes for B2C accounts (#74)
When My Account is loading the controller tries to call the GetQuotes API in CommerceRuntime to get the quotes on the account. This does not work for B2C accounts and returns a 409 validation error about the account not being B2B.
Fix this issue for B2C by conditionally calling GetQuotes when the account is B2B. Don't call the API for B2C.

There will need to be a separate fix made in Commerce to not return a 409 for this case in the future.