COM-3313 B2B P2: Support reviving expired quotes (#323)
A seller can now revive an expired quote. When editing a quote in the Expired status the Expiration Date can be changed to a future date. No other fields are editable. After editing the expiration date the seller can submit the changes and move the quote back to the In Review status.

Co-authored-by: RajeshDatir <rajesh.datir@kibocommerce.com>