COM-3683 B2B P2: Localize dates on the View/Edit Quote pages (#332)
Localize the display of dates and times on the View and Edit Quote pages. Dates and times should display in the client browser's local timezone.
This does not apply to editing the Expiration Date which is still UTC. That will be handled separately.

This was originally fixed in commit e86b324 for COM-3484. The original fix was reverted in commit 391c09d due to some issues.

Co-authored-by: RajeshDatir <rajesh.datir@kibocommerce.com>