COM-7700 - Fixed checkout page issue. (#179)
### Issue -
Some of the customers from `JellyBelly` prod tenant were not able to see the checkout page. And this was happening because while checkout we are fetching customer's active `Credits` and while fetching Credits we were using api filter `currentBalance ge 0.01` and this filter have serious performance issues. so it was causing timeouts in case if customer have lot of credits and ultimately the checkout page was erroring out.

### Resolution -
For now, i've got rid of `currentBalance` filter from api call and increased the page size from 25 to 100 (to be extra sure and also it doesn't have any impact). And did a in memory filtering to get current balances which are greater than 0.01. And this is super performant as compare to previous filtering.